+++ src/test/kotlin/logic/usecase/AuditLogUseCaseTest.kt
@@ -1,4 +1,4 @@
-package logic.useCase
+package logic.usecase
 
 import io.mockk.every
 import io.mockk.mockk
@@ -6,14 +6,12 @@ import io.mockk.verify
 import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
-import logic.exception.InvalidStateException
 import logic.exception.NoHistoryException
 import logic.repo.AuditLogRepository
 import org.damascus.logic.model.ActionType
 import org.damascus.logic.model.History.Companion.NO_STATE
 import org.damascus.logic.model.History.Companion.NO_UUID
 import org.damascus.logic.usecase.AuditLogUseCase
-import org.junit.Ignore
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
