+++ src/test/kotlin/logic/usecase/project/ProjectUseCasesTest.kt
@@ -1,4 +1,4 @@
-package logic.useCase.ProjectUseCase
+package logic.usecase.project
 
 import com.google.common.truth.Truth.assertThat
 import io.mockk.every
